TM_project/requirements.txt
Django==2.1.2
django-cleanup==0.1.9
djangorestframework==3.9.3
django-cors-headers==2.5.3